FASE 2/Evidencias de Proyecto/backend/requirements.txt
django>=5.0,<6.0
python-dotenv>=1.0
django-cors-headers>=4.4
djangorestframework>=3.15.2
pytesseract>=0.3.10
Pillow>=10.4.0
pdf2image>=1.17.0
PyPDF2>=3.0.1
